third_party/blink/web_tests/external/wpt/css/css-anchor-position/parsing/position-visibility-computed.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Property position-visibility value 'anchors-valid'
  assert_true: 'anchors-valid' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility value 'no-overflow'
  assert_true: 'no-overflow' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility value 'anchors-valid no-overflow'
  assert_true: 'anchors-valid no-overflow' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility value 'anchors-visible no-overflow'
  assert_true: 'anchors-visible no-overflow' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility value 'no-overflow anchors-valid'
  assert_true: 'no-overflow anchors-valid' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility value 'no-overflow anchors-visible'
  assert_true: 'no-overflow anchors-visible' is a supported value for position-visibility. expected true got false
[FAIL] Property position-visibility does not inherit
  assert_equals: expected "no-overflow" but got "always"
Harness: the test ran to completion.
